requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# Task queue
celery[redis]>=5.3.6
redis>=5.0.0

# Core engine
playwright>=1.41.0
selenium>=4.15.0
undetected-chromedriver>=3.5.0
openai>=1.30.0
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
setuptools>=69.0.0
